# api/routes/pipelex/pipeline.py
from __future__ import annotations

import hashlib
import hmac
from typing import TYPE_CHECKING, Annotated, Any, cast

from fastapi import APIRouter, Depends, Request
from fastapi.responses import JSONResponse
from kajson import kajson
from mthds.client.pipeline import PipelineRequest, PipelineState
from pipelex.config import get_config
from pipelex.pipe_run.delivery_assignment import DeliveryAssignment, StorageTarget, WebhookTarget
from pipelex.pipeline.pipeline_response import PipelexPipelineExecuteResponse, PipelexPipelineStartResponse
from pipelex.pipeline.pipeline_run_setup import pipeline_run_setup
from pipelex.pipeline.runner import PipelexRunner
from pipelex.system.environment import get_required_env
from pipelex.temporal.tprl_pipe.temporal_pipe_run import make_temporal_pipe_run
from pydantic import ValidationError
from typing_extensions import override

from api.error_types import ErrorType
from api.errors import ENDPOINT_HANDLED_EXCEPTIONS, raise_internal_error, raise_validation_error
from api.routes.pipelex.utils import get_current_iso_timestamp
from api.schemas.models import PipelineApiExtras

if TYPE_CHECKING:
    from mthds.models.pipe_output import VariableMultiplicity
    from mthds.models.pipeline_inputs import PipelineInputs
    from mthds.models.working_memory import WorkingMemoryAbstract
    from pipelex.core.memory.working_memory import WorkingMemory

    from api.security import RequestUser


router = APIRouter(tags=["pipeline"])

# Request bodies are parsed from the raw Request (see `_parse_request`) so structured
# inputs survive kajson decoding and callback_urls get SSRF-validated. FastAPI therefore
# can't infer the body schema from a typed parameter, so we publish it explicitly via
# `openapi_extra`. Keep these in sync with `PipelineRequest` + `PipelineApiExtras`.
_PIPELINE_BODY_PROPERTIES: dict[str, Any] = {
    "pipe_code": {
        "type": "string",
        "description": "Code of the pipe to run. Provide this and/or `mthds_contents`.",
        "example": "summarize",
    },
    "mthds_contents": {
        "type": "array",
        "items": {"type": "string"},
        "description": "Inline MTHDS bundle(s) defining the pipe(s). Always an array, even for one file.",
    },
    "inputs": {
        "type": "object",
        "description": "Pipeline inputs keyed by name. Values may be plain text, structured objects, or Document/Image references.",
        "additionalProperties": True,
    },
    "output_name": {"type": "string", "description": "Optional name of the output stuff to return."},
    "output_multiplicity": {"description": "Optional output multiplicity override (bool or int)."},
    "dynamic_output_concept_ref": {"type": "string", "description": "Optional concept ref to coerce the output to."},
}
_PIPELINE_EXECUTE_EXAMPLE: dict[str, Any] = {
    "pipe_code": "summarize",
    "mthds_contents": [
        'domain = "hello"\nmain_pipe = "summarize"\n\n[pipe.summarize]\ntype = "PipeLLM"\n'
        'description = "Summarize the input text in one sentence"\ninputs = { text = "Text" }\n'
        'output = "Text"\nprompt = "Summarize in one sentence:\\n@text"\n'
    ],
    "inputs": {"text": "Pipelex turns plain-language pipeline definitions into reproducible AI workflows."},
}
_EXECUTE_OPENAPI_EXTRA: dict[str, Any] = {
    "requestBody": {
        "required": True,
        "content": {
            "application/json": {
                "schema": {"type": "object", "properties": _PIPELINE_BODY_PROPERTIES},
                "example": _PIPELINE_EXECUTE_EXAMPLE,
            }
        },
    }
}
_START_OPENAPI_EXTRA: dict[str, Any] = {
    "requestBody": {
        "required": True,
        "content": {
            "application/json": {
                "schema": {
                    "type": "object",
                    "properties": {
                        **_PIPELINE_BODY_PROPERTIES,
                        "pipeline_run_id": {"type": "string", "description": "Optional client-supplied run id (<=128 chars)."},
                        "callback_urls": {
                            "type": "array",
                            "items": {"type": "string", "format": "uri"},
                            "description": (
                                "Optional http(s) URLs notified on completion (HMAC-signed via "
                                "X-Completion-Signature). Private/loopback/metadata hosts are rejected."
                            ),
                        },
                    },
                },
                "example": {**_PIPELINE_EXECUTE_EXAMPLE, "callback_urls": ["https://example.com/pipelex/webhook"]},
            }
        },
    }
}


def get_request_user_id(request: Request) -> str:
    """Extract the user UUID from request state (set during auth)."""
    user: RequestUser | None = getattr(request.state, "user", None)
    return user.user_id if user else "anonymous"


def _completion_signature(pipeline_run_id: str) -> str:
    """Compute the HMAC-SHA256 signature for an async pipeline completion callback.

    The signer (this server) and the verifier (your callback receiver) must
    share the same `COMPLETION_CALLBACK_SECRET`. The signature is per-run and
    the secret never travels — only the one-way hash does.
    """
    secret = get_required_env("COMPLETION_CALLBACK_SECRET")
    return hmac.new(
        secret.encode("utf-8"),
        pipeline_run_id.encode("utf-8"),
        hashlib.sha256,
    ).hexdigest()


class ApiRunner(PipelexRunner):
    """API runner that extends PipelexRunner with start_pipeline support."""

    @override
    async def start_pipeline(
        self,
        pipe_code: str | None = None,
        mthds_contents: list[str] | None = None,
        inputs: PipelineInputs | WorkingMemoryAbstract[Any] | None = None,
        output_name: str | None = None,
        output_multiplicity: VariableMultiplicity | None = None,
        dynamic_output_concept_ref: str | None = None,
        pipeline_run_id: str | None = None,
        callback_urls: list[str] | None = None,
    ) -> PipelexPipelineStartResponse:
        """Start a pipeline execution asynchronously without waiting for completion."""
        created_at = get_current_iso_timestamp()
        pipelex_inputs: PipelineInputs | WorkingMemory | None = cast("PipelineInputs | WorkingMemory | None", inputs)

        execution_config = self.execution_config or get_config().pipelex.pipeline_execution_config
        pipe_job, resolved_pipeline_run_id, _ = await pipeline_run_setup(
            execution_config=execution_config,
            library_id=self.library_id,
            library_dirs=self.library_dirs,
            pipe_code=pipe_code,
            mthds_contents=mthds_contents,
            bundle_uris=self.bundle_uris,
            inputs=pipelex_inputs,
            output_name=output_name,
            output_multiplicity=output_multiplicity,
            dynamic_output_concept_ref=dynamic_output_concept_ref,
            pipe_run_mode=self.pipe_run_mode,
            search_domain_codes=self.search_domain_codes,
            user_id=self.user_id,
            pipeline_run_id=pipeline_run_id,
        )

        delivery_assignment = DeliveryAssignment(
            storage=StorageTarget(key_prefix="results"),
            webhooks=[
                WebhookTarget(
                    url=url,
                    headers={"X-Completion-Signature": _completion_signature(resolved_pipeline_run_id)},
                )
                for url in callback_urls
            ]
            if callback_urls
            else [],
        )

        temporal_pipe_run = make_temporal_pipe_run()
        workflow_id, _handle = await temporal_pipe_run.start(
            pipe_job,
            delivery_assignment=delivery_assignment,
        )

        return PipelexPipelineStartResponse(
            pipeline_run_id=resolved_pipeline_run_id,
            created_at=created_at,
            pipeline_state=PipelineState.STARTED,
            workflow_id=workflow_id,
        )


def _decode_body(body: bytes) -> dict[str, Any]:
    """kajson-decode the body and confirm it's a dict. Raises 422 if not."""
    try:
        decoded = kajson.loads(body.decode("utf-8"))
    except (UnicodeDecodeError, ValueError) as exc:
        raise_validation_error(
            message=f"Request body is not valid JSON: {exc!s}",
            error_type=ErrorType.INVALID_JSON,
        )
    if not isinstance(decoded, dict):
        raise_validation_error(
            message="Request body must be a JSON object",
            error_type=ErrorType.INVALID_JSON,
        )
    return cast("dict[str, Any]", decoded)


def _validate_extras(request_data: dict[str, Any]) -> PipelineApiExtras:
    """Validate API-server-only fields (pipeline_run_id, callback_urls)."""
    try:
        return PipelineApiExtras.model_validate(
            {
                "pipeline_run_id": request_data.get("pipeline_run_id"),
                "callback_urls": request_data.get("callback_urls"),
            }
        )
    except ValidationError as exc:
        raise_validation_error(
            message=str(exc),
            error_type=ErrorType.INVALID_CALLBACK_URLS,
        )


async def parse_pipeline_request(request: Request) -> tuple[PipelineRequest, PipelineApiExtras]:
    """Parse and validate the request body.

    Splits the body into:
      1. The upstream `PipelineRequest` (pipe_code, mthds_contents, inputs, …)
         decoded via `kajson` so structured inputs survive without re-parsing.
      2. `PipelineApiExtras` (pipeline_run_id, callback_urls) validated by
         Pydantic — callback_urls are checked for scheme + private/loopback
         hosts to harden against SSRF.

    Body size is capped upstream by `request_body_size_middleware`.
    """
    body = await request.body()
    request_data = _decode_body(body)
    extras = _validate_extras(request_data)
    pipeline_request = PipelineRequest.from_body(request_data)
    return pipeline_request, extras


@router.post(
    "/pipeline/execute",
    response_model=PipelexPipelineExecuteResponse,
    summary="Run a pipeline and wait for the result (blocking)",
    description=(
        "Runs the pipeline and returns its full output in the response. This is the core "
        "self-hosted primitive — it has no time limit beyond what your own client/proxy "
        "allows. (Behind the hosted Pipelex gateway a ~30s synchronous cap applies; off-gateway "
        "it does not, but your own reverse proxy's read/idle timeout still does.)"
    ),
    openapi_extra=_EXECUTE_OPENAPI_EXTRA,
)
async def execute(request: Request) -> JSONResponse:
    """Execute a pipeline and wait for completion."""
    pipeline_request, _extras = await parse_pipeline_request(request)
    try:
        runner = ApiRunner(user_id=get_request_user_id(request))
        response = await runner.execute_pipeline(
            pipe_code=pipeline_request.pipe_code,
            mthds_contents=pipeline_request.mthds_contents,
            inputs=pipeline_request.inputs,
            output_name=pipeline_request.output_name,
            output_multiplicity=pipeline_request.output_multiplicity,
            dynamic_output_concept_ref=pipeline_request.dynamic_output_concept_ref,
        )
        return JSONResponse(
            content=response.model_dump(mode="json", serialize_as_any=True, by_alias=True),
        )
    except ENDPOINT_HANDLED_EXCEPTIONS as exc:
        raise_internal_error(exc, context="pipeline_execute failed")


@router.post(
    "/pipeline/start",
    response_model=PipelexPipelineStartResponse,
    summary="Start a pipeline asynchronously (fire-and-callback)",
    description=(
        "Starts the pipeline and returns a `pipeline_run_id` immediately without waiting. "
        "Optionally posts the result to `callback_urls` on completion (HMAC-signed). The runner "
        "keeps no run store, so there is no by-id polling here — durable run lifecycle "
        "(`status`/`result`/`poll`) is a Pipelex Platform feature."
    ),
    openapi_extra=_START_OPENAPI_EXTRA,
)
async def start(
    request: Request,
    parsed: Annotated[tuple[PipelineRequest, PipelineApiExtras], Depends(parse_pipeline_request)],
) -> PipelexPipelineStartResponse:
    """Start a pipeline execution asynchronously without waiting for completion."""
    pipeline_request, extras = parsed
    try:
        runner = ApiRunner(user_id=get_request_user_id(request))
        return await runner.start_pipeline(
            pipe_code=pipeline_request.pipe_code,
            mthds_contents=pipeline_request.mthds_contents,
            inputs=pipeline_request.inputs,
            output_name=pipeline_request.output_name,
            output_multiplicity=pipeline_request.output_multiplicity,
            dynamic_output_concept_ref=pipeline_request.dynamic_output_concept_ref,
            pipeline_run_id=extras.pipeline_run_id,
            callback_urls=extras.callback_urls,
        )
    except ENDPOINT_HANDLED_EXCEPTIONS as exc:
        raise_internal_error(exc, context="pipeline_start failed")