# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "BuildConceptRequest": {
        "properties": {
          "spec": {
            "additionalProperties": true,
            "description": "JSON concept specification.",
            "title": "Spec",
            "type": "object"
          }
        },
        "required": [
          "spec"
        ],
        "title": "BuildConceptRequest",
        "type": "object"
      },
      "BuildConceptResponse": {
        "properties": {
          "concept_code": {
            "description": "The concept code that was generated",
            "title": "Concept Code",
            "type": "string"
          },
          "success": {
            "default": true,
            "description": "Whether the operation was successful",
            "title": "Success",
            "type": "boolean"
          },
          "toml": {
            "description": "Generated TOML content for the concept",
            "title": "Toml",
            "type": "string"
          }
        },
        "required": [
          "concept_code",
          "toml"
        ],
        "title": "BuildConceptResponse",
        "type": "object"
      },
...
      "EdgeKind": {
        "description": "Types of edges in the execution graph.",
        "enum": [
          "control",
          "data",
          "contains",
          "selected_outcome",
          "batch_item",
          "batch_aggregate",
          "parallel_combine"
        ],
        "title": "EdgeKind",
        "type": "string"
      },
      "EdgeSpec": {
        "additionalProperties": false,
        "description": "Specification for an edge in the execution graph.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/EdgeKind"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "source_stuff_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Stuff Digest"
          },
          "target": {
            "title": "Target",
            "type": "string"
          },
          "target_stuff_digest": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Stuff Digest"
          }
        },
        "required": [
          "id",
          "source",
          "target",
          "kind"
        ],
        "title": "EdgeSpec",
        "type": "object"
      },
...
      "GraphSpec": {
        "additionalProperties": false,
        "description": "The canonical specification for a pipeline execution graph.\n\nThis is the top-level model representing a complete run graph.\nIt is versioned and designed for JSON serialization.",
        "properties": {
          "concept_registry": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Concept Registry",
            "type": "object"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "edges": {
            "items": {
              "$ref": "#/components/schemas/EdgeSpec"
            },
            "title": "Edges",
            "type": "array"
          },
          "graph_id": {
            "title": "Graph Id",
            "type": "string"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "nodes": {
            "items": {
              "$ref": "#/components/schemas/NodeSpec"
            },
            "title": "Nodes",
            "type": "array"
          },
          "pipe_registry": {
            "additionalProperties": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Pipe Registry",
            "type": "object"
          },
          "pipeline_ref": {
            "$ref": "#/components/schemas/PipelineRef"
          }
        },
        "required": [
          "graph_id",
          "created_at"
        ],
        "title": "GraphSpec",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PipeOutput": {
        "additionalProperties": false,
        "properties": {
          "graph_assembly_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Assembly Error"
          },
          "graph_spec": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GraphSpec"
              },
              {
                "type": "null"
              }
            ]
          },
          "pipeline_run_id": {
            "default": "untitled",
            "title": "Pipeline Run Id",
            "type": "string"
          },
          "working_memory": {
            "$ref": "#/components/schemas/WorkingMemory"
          },
          "working_memory_raw": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Working Memory Raw"
          }
        },
        "title": "PipeOutput",
        "type": "object"
      },
...
      "PipelexPipelineExecuteResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "main_stuff_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Main Stuff Name"
          },
          "pipe_output": {
            "$ref": "#/components/schemas/PipeOutput"
          },
          "pipeline_run_id": {
            "title": "Pipeline Run Id",
            "type": "string"
          },
          "pipeline_state": {
            "$ref": "#/components/schemas/PipelineState"
          }
        },
        "required": [
          "pipeline_run_id",
          "created_at",
          "pipeline_state",
          "pipe_output"
        ],
        "title": "PipelexPipelineExecuteResponse",
        "type": "object"
      },
      "PipelexPipelineStartResponse": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "main_stuff_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Main Stuff Name"
          },
          "pipe_output": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PipeOutput"
              },
              {
                "type": "null"
              }
            ]
          },
          "pipeline_run_id": {
            "title": "Pipeline Run Id",
            "type": "string"
          },
          "pipeline_state": {
            "$ref": "#/components/schemas/PipelineState"
          },
          "workflow_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Id"
          }
        },
        "required": [
          "pipeline_run_id",
          "created_at",
          "pipeline_state"
        ],
        "title": "PipelexPipelineStartResponse",
        "type": "object"
      },
...
      "PipelineState": {
        "enum": [
          "RUNNING",
          "COMPLETED",
          "FAILED",
          "CANCELLED",
          "ERROR",
          "STARTED"
        ],
        "title": "PipelineState",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "WorkingMemory": {
        "additionalProperties": false,
        "properties": {
          "aliases": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "object"
          },
          "root": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Stuff"
            },
            "title": "Root",
            "type": "object"
          }
        },
        "title": "WorkingMemory",
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Bearer token: the `API_KEY` (when `AUTH_MODE=api_key`) or a signed JWT (when `AUTH_MODE=jwt`). Omit when `AUTH_MODE=none`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "Pipelex",
      "url": "https://docs.pipelex.com/"
    },
    "description": "The **Pipelex Runner API** is the open-source, self-hostable engine that executes\nPipelex pipelines (MTHDS bundles) over HTTP. It is **stateless**: it needs no\ndatabase, queue, or object store to run pipelines \u2014 boot the Docker image, give it\nan inference backend, and call `/api/v1/pipeline/execute`.\n\n### What you must provide\n- **Inference backend** \u2014 by default a `PIPELEX_GATEWAY_API_KEY` (the Pipelex Gateway\n  routes to every supported model with one credential). To call providers directly\n  (OpenAI, Anthropic, Bedrock, Vertex, \u2026) you reconfigure the mounted `.pipelex/`\n  config \u2014 not just an env var. See the Configuration docs.\n- **Auth mode** (`AUTH_MODE`) \u2014 `none` (default, local), `api_key` (`API_KEY`), or\n  `jwt` (`JWT_SECRET_KEY`). The caller is identified by the credential; the runner\n  never accepts a user id as a request argument.\n\n### What is NOT in this runner\nDurable run storage and by-id polling (`start`/`status`/`result`/`poll`), user-scoped\nfile storage (`/upload`, `/resolve-storage-url`), org/billing \u2014 these are **Pipelex\nPlatform** features. The open-source runner is execution-only; run a pipeline and get\nits output back synchronously via `/pipeline/execute`.\n",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/license/mit"
    },
...
    "version": "0.1.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/build/concept": {
      "post": {
        "description": "Convert a JSON concept spec to TOML format.",
        "operationId": "build_concept_api_v1_build_concept_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildConceptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildConceptResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Convert a JSON concept spec to TOML",
        "tags": [
          "agent"
        ]
      }
    },
...
    "/api/v1/pipeline/execute": {
      "post": {
        "description": "Runs the pipeline and returns its full output in the response. This is the core self-hosted primitive \u2014 it has no time limit beyond what your own client/proxy allows. (Behind the hosted Pipelex gateway a ~30s synchronous cap applies; off-gateway it does not, but your own reverse proxy's read/idle timeout still does.)",
        "operationId": "execute_api_v1_pipeline_execute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "inputs": {
                  "text": "Pipelex turns plain-language pipeline definitions into reproducible AI workflows."
                },
                "mthds_contents": [
                  "domain = \"hello\"\nmain_pipe = \"summarize\"\n\n[pipe.summarize]\ntype = \"PipeLLM\"\ndescription = \"Summarize the input text in one sentence\"\ninputs = { text = \"Text\" }\noutput = \"Text\"\nprompt = \"Summarize in one sentence:\\n@text\"\n"
                ],
                "pipe_code": "summarize"
              },
              "schema": {
                "properties": {
                  "dynamic_output_concept_ref": {
                    "description": "Optional concept ref to coerce the output to.",
                    "type": "string"
                  },
                  "inputs": {
                    "additionalProperties": true,
                    "description": "Pipeline inputs keyed by name. Values may be plain text, structured objects, or Document/Image references.",
                    "type": "object"
                  },
                  "mthds_contents": {
                    "description": "Inline MTHDS bundle(s) defining the pipe(s). Always an array, even for one file.",
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "output_multiplicity": {
                    "description": "Optional output multiplicity override (bool or int)."
                  },
                  "output_name": {
                    "description": "Optional name of the output stuff to return.",
                    "type": "string"
                  },
                  "pipe_code": {
                    "description": "Code of the pipe to run. Provide this and/or `mthds_contents`.",
                    "example": "summarize",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelexPipelineExecuteResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Run a pipeline and wait for the result (blocking)",
        "tags": [
          "pipeline"
        ]
      }
    },
    "/api/v1/pipeline/start": {
      "post": {
        "description": "Starts the pipeline and returns a `pipeline_run_id` immediately without waiting. Optionally posts the result to `callback_urls` on completion (HMAC-signed). The runner keeps no run store, so there is no by-id polling here \u2014 durable run lifecycle (`status`/`result`/`poll`) is a Pipelex Platform feature.",
        "operationId": "start_api_v1_pipeline_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "example": {
                "callback_urls": [
                  "https://example.com/pipelex/webhook"
                ],
                "inputs": {
                  "text": "Pipelex turns plain-language pipeline definitions into reproducible AI workflows."
                },
                "mthds_contents": [
                  "domain = \"hello\"\nmain_pipe = \"summarize\"\n\n[pipe.summarize]\ntype = \"PipeLLM\"\ndescription = \"Summarize the input text in one sentence\"\ninputs = { text = \"Text\" }\noutput = \"Text\"\nprompt = \"Summarize in one sentence:\\n@text\"\n"
                ],
                "pipe_code": "summarize"
              },
              "schema": {
                "properties": {
                  "callback_urls": {
                    "description": "Optional http(s) URLs notified on completion (HMAC-signed via X-Completion-Signature). Private/loopback/metadata hosts are rejected.",
                    "items": {
                      "format": "uri",
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "dynamic_output_concept_ref": {
                    "description": "Optional concept ref to coerce the output to.",
                    "type": "string"
                  },
                  "inputs": {
                    "additionalProperties": true,
                    "description": "Pipeline inputs keyed by name. Values may be plain text, structured objects, or Document/Image references.",
                    "type": "object"
                  },
                  "mthds_contents": {
                    "description": "Inline MTHDS bundle(s) defining the pipe(s). Always an array, even for one file.",
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "output_multiplicity": {
                    "description": "Optional output multiplicity override (bool or int)."
                  },
                  "output_name": {
                    "description": "Optional name of the output stuff to return.",
                    "type": "string"
                  },
                  "pipe_code": {
                    "description": "Code of the pipe to run. Provide this and/or `mthds_contents`.",
                    "example": "summarize",
                    "type": "string"
                  },
                  "pipeline_run_id": {
                    "description": "Optional client-supplied run id (<=128 chars).",
                    "type": "string"
                  }
                },
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelexPipelineStartResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Start a pipeline asynchronously (fire-and-callback)",
        "tags": [
          "pipeline"
        ]
      }
    },
...
  },
  "security": [
    {},
    {
      "BearerAuth": []
    }
  ],
  "servers": [
...
}